feat(lunox-event-typeorm): add queue job and failed job models
