+++ packages/lunox-event-typeorm/mysql.d.ts
@@ -1,0 +1,1 @@
+export * from "./dist/models/mysql"
