+++ packages/lunox-event-typeorm/postgre.d.ts
@@ -1,0 +1,1 @@
+export * from "./dist/models/postgre"
