+++ packages/lunox-event-typeorm/sqlite.d.ts
@@ -1,0 +1,1 @@
+export * from "./dist/models/sqlite"
