+++ packages/lunox-event-typeorm/src/models/mysql/index.ts
@@ -1,0 +1,7 @@
+import QueueJobFailed from "./QueueJobFailed";
+import QueueJob from "./QueueJob";
+
+export {
+  QueueJobFailed,
+  QueueJob,
+};
