+++ packages/lunox-event-typeorm/src/models/postgre/QueueJobFailed.ts
@@ -1,22 +1,23 @@
+import { QueueJobFailedSchema } from "@lunoxjs/event/contracts";
 import { Column, Entity, PrimaryGeneratedColumn } from "typeorm";
 
 @Entity(
-  "queue_failed_jobs"
+  "queue_failed_jobs",
 )
-class QueueJobFailedSqlite {
+class QueueJobFailed implements QueueJobFailedSchema {
   @PrimaryGeneratedColumn()
   id!: number;
 
   @Column("varchar")
   queue!: string;
 
-  @Column("blob")
+  @Column("bytea")
   payload!: Buffer;
 
   @Column("text")
   exception!: string;
 
-  @Column("datetime", { nullable: true })
+  @Column("timestamptz", { nullable: true })
   failed_at?: Date;
 }
-export default QueueJobFailedSqlite;
+export default QueueJobFailed;
