+++ packages/lunox-event-typeorm/tsup.config.ts
@@ -1,6 +1,6 @@
 import { defineConfig } from "tsup";
 export default defineConfig({
-  entry: ["src/index.ts"],
+  entry: ["src/index.ts", "src/models/*/index.ts"],
   format: "esm",
   experimentalDts: true,
   clean: true,
