+++ packages/lunox-mail/package.json
@@ -32,6 +32,7 @@
   "devDependencies": {
     "@lunoxjs/core": "workspace:*",
     "@lunoxjs/event": "workspace:*",
+    "@lunoxjs/event-typeorm": "workspace:2.0.0-beta.3.5.0",
     "@lunoxjs/test": "workspace:*",
     "@lunoxjs/typeorm": "workspace:*",
     "@lunoxjs/view": "workspace:*",
