+++ packages/lunox-mail/test/config/database.ts
@@ -1,6 +1,5 @@
 import type { DatabaseConfig } from "@lunoxjs/typeorm/contracts";
-import QueueJobMysql from "../models/typeorm/QueueJobMysql";
-import QueueJobFailedMysql from "../models/typeorm/QueueJobFailedMysql";
+import {QueueJob, QueueJobFailed} from "@lunoxjs/event-typeorm/mysql";
 
 export default {
   defaultConnection: env("DB_CONNECTION", "mysql"),
@@ -8,7 +7,7 @@ export default {
   connections: {
     mysql: {
       type: "mysql",
-      entities: [QueueJobMysql, QueueJobFailedMysql],
+      entities: [QueueJob, QueueJobFailed],
       host: env("DB_HOST", "localhost"),
       port: env("DB_PORT", "3306"),
       username: env("DB_USERNAME"),
