--- packages/lunox-mail/test/models
@@ -1,1 +1,0 @@
-/home/dmuggle/GIT/lunox/packages/lunox-event/test/app/models
